snapshot of project "ncurses", label v5_5_20051203

+++ ncurses/curses.priv.h
@@ -34,7 +34,7 @@
 
 
 /*
- * $Id: curses.priv.h,v 1.287 2005/11/20 01:32:15 tom Exp $
+ * $Id: curses.priv.h,v 1.288 2005/11/26 20:07:45 tom Exp $
  *
  *	curses.priv.h
  *
@@ -1092,7 +1092,7 @@ extern NCURSES_EXPORT(char *) _nc_printf_string (const char *, va_list);
 
 /* tries.c */
 extern NCURSES_EXPORT(void) _nc_add_to_try (struct tries **, const char *, unsigned);
-extern NCURSES_EXPORT(char *) _nc_expand_try (struct tries *, unsigned short, int *, size_t);
+extern NCURSES_EXPORT(char *) _nc_expand_try (struct tries *, unsigned, int *, size_t);
 extern NCURSES_EXPORT(int) _nc_remove_key (struct tries **, unsigned);
 extern NCURSES_EXPORT(int) _nc_remove_string (struct tries **, const char *);
 
@@ -1111,7 +1111,7 @@ extern NCURSES_EXPORT(int) _nc_ospeed (int);
 extern NCURSES_EXPORT(int) _nc_outch (int);
 extern NCURSES_EXPORT(int) _nc_setupscreen (int, int const, FILE *);
 extern NCURSES_EXPORT(int) _nc_timed_wait(int, int, int * EVENTLIST_2nd(_nc_eventlist *));
-extern NCURSES_EXPORT(void) _nc_do_color (int, int, bool, int (*)(int));
+extern NCURSES_EXPORT(void) _nc_do_color (short, short, bool, int (*)(int));
 extern NCURSES_EXPORT(void) _nc_flush (void);
 extern NCURSES_EXPORT(void) _nc_free_entry(ENTRY *, TERMTYPE *);
 extern NCURSES_EXPORT(void) _nc_freeall (void);
